+++ android/src/main/java/com/sean/rao/ali_auth/config/DialogPortConfig.java
@@ -7,6 +7,7 @@ import androidx.core.graphics.drawable.RoundedBitmapDrawable;
 import androidx.core.graphics.drawable.RoundedBitmapDrawableFactory;
 
 import com.mobile.auth.gatewayauth.AuthRegisterViewConfig;
+import com.sean.rao.ali_auth.common.CustomAuthUIControlClickListener;
 import com.sean.rao.ali_auth.utils.AppUtils;
 import com.sean.rao.ali_auth.utils.UtilTool;
 
@@ -56,6 +57,7 @@ public class DialogPortConfig extends BaseUIConfig {
                 showResult("500000", "背景处理时出现错误", e.getMessage());
             }
         }
+        mAuthHelper.setUIClickListener(new CustomAuthUIControlClickListener());
         mAuthHelper.setAuthUIConfig(config.setScreenOrientation(authPageOrientation).create());
     }
 }